+++ src/observable/observable_comp.rs
@@ -34,7 +34,7 @@ pub trait SubscribeComplete<N, C> {
     self,
     next: N,
     complete: C,
-  ) -> SubscriptionGuard<Self::Unsub>;
+  ) -> SubscriptionWrapper<Self::Unsub>;
 }
 
 impl<'a, S, N, C> SubscribeComplete<N, C> for S
@@ -48,13 +48,13 @@ where
     self,
     next: N,
     complete: C,
-  ) -> SubscriptionGuard<Self::Unsub>
+  ) -> SubscriptionWrapper<Self::Unsub>
   where
     Self: Sized,
   {
     let unsub =
       self.actual_subscribe(Subscriber::local(ObserverComp { next, complete }));
-    SubscriptionGuard(unsub)
+    SubscriptionWrapper(unsub)
   }
 }
 
@@ -69,14 +69,14 @@ where
     self,
     next: N,
     complete: C,
-  ) -> SubscriptionGuard<Self::Unsub>
+  ) -> SubscriptionWrapper<Self::Unsub>
   where
     Self: Sized,
   {
     let unsub = self
       .0
       .actual_subscribe(Subscriber::shared(ObserverComp { next, complete }));
-    SubscriptionGuard(unsub)
+    SubscriptionWrapper(unsub)
   }
 }
 
@@ -85,7 +85,10 @@ fn raii() {
   let mut times = 0;
   {
     let mut subject = Subject::local();
-    subject.clone().subscribe_complete(|_| times += 1, || {});
+    subject
+      .fork()
+      .subscribe_complete(|_| times += 1, || {})
+      .unsubscribe_when_dropped();
     subject.next(());
   }
   assert_eq!(times, 0);