+++ src/subscription.rs
@@ -187,6 +187,23 @@ impl<Item, Err> SubscriptionLike
   subscription_direct_impl_proxy!();
 }
 
+/// Wrapper around a subscription which provides the `unsubscribe_when_dropped()` method.
+pub struct SubscriptionWrapper<T: SubscriptionLike>(pub(crate) T);
+
+impl<T: SubscriptionLike> SubscriptionWrapper<T> {
+
+  /// Activates "RAII" behavior for this subscription. That means `unsubscribe()` will be called
+  /// automatically as soon as the returned value goes out of scope.
+  ///
+  /// **Attention:** If you don't assign the return value to a variable, `unsubscribe()` is called
+  /// immediately, which is probably not what you want!
+  pub fn unsubscribe_when_dropped(self) -> SubscriptionGuard<T> {
+    SubscriptionGuard(self.0)
+  }
+}
+
+subscription_proxy_impl!(SubscriptionWrapper<T>, { 0 }, T);
+
 /// An RAII implementation of a "scoped subscribed" of a subscription.
 /// When this structure is dropped (falls out of scope), the subscription will
 /// be unsubscribed.
@@ -195,8 +212,6 @@ impl<T: SubscriptionLike> Drop for SubscriptionGuard<T> {
   #[inline]
   fn drop(&mut self) { self.0.unsubscribe() }
 }
-
-subscription_proxy_impl!(SubscriptionGuard<T>, { 0 }, T);
 
 #[cfg(test)]
 mod test {